Legacy classes cleanup
