+++ modules/itc/src/main/scala/lucuma/itc/legacy/encoders.scala
@@ -22,7 +22,7 @@ import lucuma.itc.ItcSeries
 import lucuma.itc.SeriesDataType
 import lucuma.itc.search.ObservingMode.Spectroscopy._
 import lucuma.itc.search.*
-import lucuma.itc.search.syntax.all.*
+import lucuma.itc.legacy.syntax.all.*
 import lucuma.itc.syntax.all.given
 
 import java.math.MathContext
